refactor: 인터페이스 검증 디폴트 메서드 제거
- 중복 제거 효과가 크지 않으며, 검증이 구현체의 책임이라고 판단하여 디폴드 메서드를 제거합니다.
